+++ ledger/src/blockstore_meta.rs
@@ -623,12 +623,6 @@ impl DuplicateSlotProof {
 }
 
 #[derive(Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
-pub struct TransactionStatusIndexMeta {
-    pub max_slot: Slot,
-    pub frozen: bool,
-}
-
-#[derive(Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
 pub struct AddressSignatureMeta {
     pub writeable: bool,
 }